Parsing interface 생성 및 Parsing -> ParsingImpl
