+++ Calculation/src/test/java/com/programmers/java/calculation/CalcualateTest.java
@@ -2,7 +2,7 @@ package com.programmers.java.calculation;
 
 import com.programmers.java.calculation.calculate.Calculate;
 import com.programmers.java.calculation.calculate.CalculateBasicImpl;
-import com.programmers.java.calculation.parse.Parsing;
+import com.programmers.java.calculation.parse.ParsingImpl;
 import com.programmers.java.calculation.parse.ValidationImpl;
 import com.programmers.java.calculation.parse.Validation;
 import org.junit.jupiter.api.Test;
@@ -11,7 +11,7 @@ import java.util.List;
 
 public class CalcualateTest {
 
-    Parsing parsing = new Parsing();
+    ParsingImpl parsing = new ParsingImpl();
     Validation validation = new ValidationImpl();
     Calculate calculate = new CalculateBasicImpl();
